Print hex value of magicNumber instead of decimal Fix a bug in OnDiskByteSink.writeList where HAS_VALUE flag was not written in the file Slightly change BasePacket and BaseResponse (have toByteSink accept byteSink as a parameter instead of creating it inside of it) Make Packet and Response tests run on both InMemoryByteSink and OnDiskByteSink

+++ chat-common/src/main/java/core/exception/MaxListSizeExceededException.kt
@@ -2,4 +2,7 @@ package core.exception
 
 import java.lang.Exception
 
-class MaxListSizeExceededException(listSize: Short, maxListSize: Int) : Exception("Exceeded max list size (listSize: $listSize, maxListSize: $maxListSize)")
+class MaxListSizeExceededException(
+  listSize: Int,
+  maxListSize: Int
+) : Exception("Exceeded max list size (listSize: $listSize, maxListSize: $maxListSize)")
